third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-blocked-different-suborigins-expected.txt


This is a testharness.js-based test.
FAIL Block a frame in one suborigin from accessing another suborigin Uncaught Error: assert_equals: The parent frame should not be able to get the secret value from the child iframe. expected (string) "" but got (undefined) undefined
Harness: the test ran to completion.
